nepenthes.submit-gotek: removed ugly `could not delete submitted file ..' log messages if spooling was disabled anyway

+++ modules/submit-gotek/gotekDATADialogue.cpp
@@ -272,7 +272,7 @@ ConsumeLevel gotekDATADialogue::connectionLost(Message *msg)
  */
 ConsumeLevel gotekDATADialogue::connectionShutdown(Message *msg)
 {
-	if(m_State == GDATA_DONE)
+	if(m_State == GDATA_DONE && !m_GotekContext->m_FileName.empty())
 	{
 		if(unlink(m_GotekContext->m_FileName.c_str()) < 0)
 		{
